refactor: simplify Logger structure and put it behind internal namespace
